Adding some fina l touches

+++ PHP/accountdelete.inc.php
@@ -1,34 +1,39 @@
 <?php
+if ($_SERVER["REQUEST_METHOD"] == "POST")
+{
+    $username = $_POST["username"];//replace with form variable
+    $pwd= $_POST["pwd"];//replace with form variable
+    $accountid = $_POST["accountID"];
 
     try 
     {
         require_once("dbapi.inc.php");//links file connects to the database
 
-        $json_data = file_get_contents("deleteaccount.json");
-        $deleteaccounts = json_decode($json_data, JSON_OBJECT_AS_ARRAY);//retries json file data turn it into a string an place it in an array
-
-        $query = "DELETE FROM username = :username AND pwd = :pwd;";//deletes user
-
-        foreach($deleteaccounts as $deleteaccount)//goes through the json file retrieved
-        {
-            $username = $deleteaccount["username"];//sorts data from json file to appropriate variables
-            $pwd = $deleteaccount["pwd"];
-        }
+        $query = "DELETE FROM username = $username AND pwd = $pwd AND acountID = $accountid;";//deletes user
 
         $statement = $pdo->prepare($query);
-        
-        $statement->bindParam(":username", $username);
-        $statement->bindParam(":pwd", $pwd);
+
+        $statement->execute();//submit data from user
 
         $pdo = null;//closing of connection to database
         $statement = null;
 
+        header("mainpagesearch.inc.php");
         die();
-        //header("");makes sure the user logins into there own page
+      
         
     } 
     catch (PDOException $e) 
     {
         die("Failed". $e->getMessage());//it it fails it just terminates the script
     }
+    
+   
+
+}
+else
+{
+    header("mainpagesearch.inc.php");
+
+}
